+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     - Shared Token Authenticator uses bearer scheme and multiple tokens
     - Fix REST Server middleware stack
     - Added Response Time middleware
+    - Added agent and response length to HTTP request logging
 
 - 0.2.0-beta
     - Update to Rubix ML 0.2.0